auxclick/auxclick_event-manual.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>Clicking with primary vs non-primary buttons</title>
    <link rel="help" href="https://wicg.github.io/auxclick/">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <style>
#target {
  background-color: green;
  height: 200px;
  width: 200px;
}
    </style>
  </head>
  <body>
    <h1>Clicking with primary vs non-primary buttons</h1>
    <p>Double-click on the green box with a non-primary button. When using mouse any button other than the left button is non-primary. If a "PASS" result appears, the test passes; otherwise, it fails.</p>
    <div id="target"></div>
    <script>
    var test_auxclick = async_test("auxclick event received.");
    var target = document.querySelector('#target');
    document.addEventListener('contextmenu', event => { event.preventDefault(); });
    ['click', 'dblclick'].forEach(eventName => {
        target.addEventListener(eventName, () => {
            test_auxclick.step(() => {
                assert_unreached(eventName + ' event should not be dispatched for non-primary buttons.');
            });
        });
        document.addEventListener(eventName, () => {
            test_auxclick.step(() => {
                assert_unreached('document should not receive ' + eventName + ' for non-primary buttons.');
            });
        }, true);
    });
    var click_count = 0;
    target.addEventListener('auxclick', event => {
        test_auxclick.step(() => {
            assert_equals(event.detail, ++click_count, 'detail attribute of auxclick should be the click count.');
        });
        if (click_count == 2)
            test_auxclick.done();
    });
    </script>
  </body>
</html>
